refactor(builder): extract marking-scheme upload into useMarkingScheme hook
Moves the ~290 lines of marking-scheme logic (text/vision extraction, parse
request, detection, fill + append, highlight) out of the 11k-line CourseBuilder
into src/.../hooks/use-marking-scheme.ts. The gnarly match/append core is now a
pure, unit-tested applySchemeMatches() in the shared lib (3 new cases). Behaviour
preserving — CourseBuilder calls the hook and consumes its returns unchanged.

First step of the DMI/PCI/marking-scheme extraction; usePci / useDmiEditor follow.
