+++ tutorme-app/public/sw.js
@@ -1,4 +1,4 @@
-"use strict";(()=>{var u="tutorme-app",l="1782712486007",c={PRECACHE:`${u}-precache-${l}`,RUNTIME:`${u}-runtime-${l}`,OFFLINE:`${u}-offline-${l}`,IMAGES:`${u}-images-${l}`,API:`${u}-api-${l}`,STATIC:`${u}-static-${l}`},p=100,E=300,q=3600*1,S=1e4,A=1e4;var R=["/offline.html","/manifest.json"];self.addEventListener("install",e=>{console.warn("[Service Worker] Installing...",l),self.skipWaiting(),e.waitUntil(caches.open(c.PRECACHE).then(t=>t.addAll(R)).catch(t=>console.warn("[Service Worker] Precache failed:",t)))});self.addEventListener("activate",e=>{console.warn("[Service Worker] Activating...",l),e.waitUntil(caches.keys().then(t=>Promise.all(t.filter(s=>s.startsWith(`${u}-`)&&!Object.values(c).includes(s)).map(s=>caches.delete(s)))).then(()=>self.clients.claim()).catch(t=>(console.error("[Service Worker] Activation failed:",t),self.clients.claim())))});async function b(e){try{let t=await caches.open(c.OFFLINE),s=await t.match("offline-queue"),n={requests:[],lastSync:Date.now()};if(s)try{n=await s.json()}catch(o){}n.requests.push(e),n.lastSync=Date.now(),n.requests.length>p&&(n.requests=n.requests.slice(-p)),await t.put("offline-queue",new Response(JSON.stringify(n),{headers:{"Content-Type":"application/json"}}))}catch(t){console.error("[Service Worker] Failed to store offline request:",t)}}async function g(){try{let t=await(await caches.open(c.OFFLINE)).match("offline-queue");if(t){let s=await t.json();return Array.isArray(s.requests)?s.requests:[]}}catch(e){console.error("[Service Worker] Failed to get offline requests:",e)}return[]}async function C(){try{await(await caches.open(c.OFFLINE)).delete("offline-queue")}catch(e){console.error("[Service Worker] Failed to clear offline requests:",e)}}function w(e,t){let s=e.headers.get("date");if(!s)return!1;let n=(Date.now()-new Date(s).getTime())/1e3;return isNaN(n)?!1:n<t}self.addEventListener("fetch",e=>{var o;let{request:t}=e,s=t.method==="GET",n;try{n=new URL(t.url)}catch(a){return}if(t.mode==="navigate"||s&&((o=t.headers.get("accept"))!=null&&o.includes("text/html"))){e.respondWith(I(e));return}if(n.pathname.startsWith("/api/")){t.method==="GET"&&e.respondWith(O(e));return}if(s){if(/\.(?:png|jpg|jpeg|svg|gif|webp|ico)$/i.test(n.pathname)){e.respondWith(T(e));return}if(/\.(?:js|css|woff2|ttf|json)$/i.test(n.pathname)){e.respondWith(m(e));return}if(n.hostname.includes("fonts.googleapis.com")||n.hostname.includes("fonts.gstatic.com")){e.respondWith(m(e));return}}});async function I(e){let t=new AbortController,s=setTimeout(()=>t.abort(),A);try{let n=await fetch(e.request,{signal:t.signal});return clearTimeout(s),n}catch(n){return clearTimeout(s),await(await caches.open(c.PRECACHE)).match("/offline.html")||new Response(_(),{headers:{"Content-Type":"text/html; charset=utf-8","Cache-Control":"no-cache","X-Offline":"true"}})}}async function O(e){try{let t=new AbortController,s=setTimeout(()=>t.abort(),S),n=await fetch(e.request.clone(),{signal:t.signal});if(clearTimeout(s),n.ok){let o=await caches.open(c.API),a=n.clone();await o.put(e.request,a)}return n}catch(t){console.error("[SW] API fetch failed:",e.request.url,e.request.method,t==null?void 0:t.name,t==null?void 0:t.message);let n=await(await caches.open(c.API)).match(e.request);if(n&&w(n,E))return n;let o=e.request.method;if(o!=="GET"&&o!=="HEAD"){let a={id:Date.now().toString(),method:o,url:e.request.url,headers:Object.fromEntries(e.request.headers.entries()),body:await e.request.text(),timestamp:Date.now()};await b(a)}return new Response(JSON.stringify({error:"offline",message:"You are offline. Request will sync when connected."}),{status:503,headers:{"Content-Type":"application/json","X-Offline":"true"}})}}async function T(e){let t=await caches.open(c.IMAGES),s=await t.match(e.request);if(s)return s;try{let n=await fetch(e.request);return n.ok&&n.status===200&&await t.put(e.request,n.clone()),n}catch(n){return new Response("",{status:503})}}function h(e,t){let s=e.headers.get("content-type")||"";return s.includes(t)?!0:(console.warn("[SW] Invalid content-type for static:",e.url,"got:",s),!1)}async function m(e){let t=new URL(e.request.url),s=t.pathname.endsWith(".js"),n=t.pathname.endsWith(".css"),o=s?"javascript":n?"css":"",a=await caches.open(c.STATIC),r=await a.match(e.request);if(r)if(o&&!h(r,o))console.warn("[SW] Evicting corrupted cached static:",e.request.url),await a.delete(e.request);else return w(r,q)||fetch(e.request).then(async f=>{f.ok&&f.status===200&&(!o||h(f,o))&&await a.put(e.request,f.clone())}).catch(()=>{}),r;try{let i=await fetch(e.request);return i.ok&&i.status===200&&(!o||h(i,o))&&await a.put(e.request,i.clone()),i}catch(i){return new Response("",{status:503})}}function _(){return`<!DOCTYPE html>
+"use strict";(()=>{var u="tutorme-app",l="1782718157527",c={PRECACHE:`${u}-precache-${l}`,RUNTIME:`${u}-runtime-${l}`,OFFLINE:`${u}-offline-${l}`,IMAGES:`${u}-images-${l}`,API:`${u}-api-${l}`,STATIC:`${u}-static-${l}`},p=100,E=300,q=3600*1,S=1e4,A=1e4;var R=["/offline.html","/manifest.json"];self.addEventListener("install",e=>{console.warn("[Service Worker] Installing...",l),self.skipWaiting(),e.waitUntil(caches.open(c.PRECACHE).then(t=>t.addAll(R)).catch(t=>console.warn("[Service Worker] Precache failed:",t)))});self.addEventListener("activate",e=>{console.warn("[Service Worker] Activating...",l),e.waitUntil(caches.keys().then(t=>Promise.all(t.filter(s=>s.startsWith(`${u}-`)&&!Object.values(c).includes(s)).map(s=>caches.delete(s)))).then(()=>self.clients.claim()).catch(t=>(console.error("[Service Worker] Activation failed:",t),self.clients.claim())))});async function b(e){try{let t=await caches.open(c.OFFLINE),s=await t.match("offline-queue"),n={requests:[],lastSync:Date.now()};if(s)try{n=await s.json()}catch(o){}n.requests.push(e),n.lastSync=Date.now(),n.requests.length>p&&(n.requests=n.requests.slice(-p)),await t.put("offline-queue",new Response(JSON.stringify(n),{headers:{"Content-Type":"application/json"}}))}catch(t){console.error("[Service Worker] Failed to store offline request:",t)}}async function g(){try{let t=await(await caches.open(c.OFFLINE)).match("offline-queue");if(t){let s=await t.json();return Array.isArray(s.requests)?s.requests:[]}}catch(e){console.error("[Service Worker] Failed to get offline requests:",e)}return[]}async function C(){try{await(await caches.open(c.OFFLINE)).delete("offline-queue")}catch(e){console.error("[Service Worker] Failed to clear offline requests:",e)}}function w(e,t){let s=e.headers.get("date");if(!s)return!1;let n=(Date.now()-new Date(s).getTime())/1e3;return isNaN(n)?!1:n<t}self.addEventListener("fetch",e=>{var o;let{request:t}=e,s=t.method==="GET",n;try{n=new URL(t.url)}catch(a){return}if(t.mode==="navigate"||s&&((o=t.headers.get("accept"))!=null&&o.includes("text/html"))){e.respondWith(I(e));return}if(n.pathname.startsWith("/api/")){t.method==="GET"&&e.respondWith(O(e));return}if(s){if(/\.(?:png|jpg|jpeg|svg|gif|webp|ico)$/i.test(n.pathname)){e.respondWith(T(e));return}if(/\.(?:js|css|woff2|ttf|json)$/i.test(n.pathname)){e.respondWith(m(e));return}if(n.hostname.includes("fonts.googleapis.com")||n.hostname.includes("fonts.gstatic.com")){e.respondWith(m(e));return}}});async function I(e){let t=new AbortController,s=setTimeout(()=>t.abort(),A);try{let n=await fetch(e.request,{signal:t.signal});return clearTimeout(s),n}catch(n){return clearTimeout(s),await(await caches.open(c.PRECACHE)).match("/offline.html")||new Response(_(),{headers:{"Content-Type":"text/html; charset=utf-8","Cache-Control":"no-cache","X-Offline":"true"}})}}async function O(e){try{let t=new AbortController,s=setTimeout(()=>t.abort(),S),n=await fetch(e.request.clone(),{signal:t.signal});if(clearTimeout(s),n.ok){let o=await caches.open(c.API),a=n.clone();await o.put(e.request,a)}return n}catch(t){console.error("[SW] API fetch failed:",e.request.url,e.request.method,t==null?void 0:t.name,t==null?void 0:t.message);let n=await(await caches.open(c.API)).match(e.request);if(n&&w(n,E))return n;let o=e.request.method;if(o!=="GET"&&o!=="HEAD"){let a={id:Date.now().toString(),method:o,url:e.request.url,headers:Object.fromEntries(e.request.headers.entries()),body:await e.request.text(),timestamp:Date.now()};await b(a)}return new Response(JSON.stringify({error:"offline",message:"You are offline. Request will sync when connected."}),{status:503,headers:{"Content-Type":"application/json","X-Offline":"true"}})}}async function T(e){let t=await caches.open(c.IMAGES),s=await t.match(e.request);if(s)return s;try{let n=await fetch(e.request);return n.ok&&n.status===200&&await t.put(e.request,n.clone()),n}catch(n){return new Response("",{status:503})}}function h(e,t){let s=e.headers.get("content-type")||"";return s.includes(t)?!0:(console.warn("[SW] Invalid content-type for static:",e.url,"got:",s),!1)}async function m(e){let t=new URL(e.request.url),s=t.pathname.endsWith(".js"),n=t.pathname.endsWith(".css"),o=s?"javascript":n?"css":"",a=await caches.open(c.STATIC),r=await a.match(e.request);if(r)if(o&&!h(r,o))console.warn("[SW] Evicting corrupted cached static:",e.request.url),await a.delete(e.request);else return w(r,q)||fetch(e.request).then(async f=>{f.ok&&f.status===200&&(!o||h(f,o))&&await a.put(e.request,f.clone())}).catch(()=>{}),r;try{let i=await fetch(e.request);return i.ok&&i.status===200&&(!o||h(i,o))&&await a.put(e.request,i.clone()),i}catch(i){return new Response("",{status:503})}}function _(){return`<!DOCTYPE html>
 <html lang="en">
 <head>
   <meta charset="utf-8">
