+++ tutorme-app/src/app/[locale]/tutor/dashboard/components/CourseBuilder.tsx
@@ -116,12 +116,8 @@ import {
   DMI_QUESTION_TYPE_LABELS,
   type DmiQuestionType,
 } from '@/lib/assessment/question-types'
-import {
-  deriveExamContext,
-  EXAM_BOARDS,
-  refKey,
-  extractQuestionRef,
-} from '@/lib/assessment/marking-scheme'
+import { deriveExamContext, EXAM_BOARDS, extractQuestionRef } from '@/lib/assessment/marking-scheme'
+import { useMarkingScheme } from './hooks/use-marking-scheme'
 import { Separator } from '@/components/ui/separator'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { ResizablePanelGroup, ResizablePanel, ResizableHandle } from '@/components/ui/resizable'
@@ -1110,12 +1106,6 @@ export const CourseBuilder = forwardRef<CourseBuilderRef, CourseBuilderProps>(
     const [dmiEditor, setDmiEditor] = useState<{ source: 'task' | 'assessment' } | null>(null)
     // "Upload marking scheme": parse an uploaded scheme and fill each question's
     // answer/rubric by matching question numbers.
-    const [markingSchemeLoading, setMarkingSchemeLoading] = useState(false)
-    const markingSchemeInputRef = useRef<HTMLInputElement | null>(null)
-    // IDs of rows just appended from a marking scheme, briefly highlighted so the
-    // tutor can spot the new questions among the existing ones.
-    const [recentlyAddedRowIds, setRecentlyAddedRowIds] = useState<Set<string>>(new Set())
-    const dmiRowsRef = useRef<HTMLDivElement | null>(null)
     // Tutor's answer-reveal policy applied to deploys: when students may see the
     // correct answers. Default 'instant' preserves the existing live-feedback
     // behaviour; the tutor can switch to reveal-after-submit or hidden.
@@ -3028,294 +3018,29 @@ FEEDBACK: [one or two short sentences explaining the score]`
     // Whether the badge's inline board/subject editor is open.
     const [editingExamContext, setEditingExamContext] = useState(false)
 
-    // Read text from an uploaded marking scheme (PDF via pdfjs, or plain text).
-    const extractMarkingSchemeText = async (file: File): Promise<string> => {
-      const isPdf = file.type === 'application/pdf' || file.name.toLowerCase().endsWith('.pdf')
-      if (!isPdf) {
-        try {
-          return await file.text()
-        } catch {
-          return ''
-        }
-      }
-      try {
-        const arrayBuffer = await file.arrayBuffer()
-        const pdfjs = await import('pdfjs-dist')
-        if (typeof window !== 'undefined') {
-          const opts = (pdfjs as { GlobalWorkerOptions?: { workerSrc?: string } })
-            .GlobalWorkerOptions
-          if (opts && !opts.workerSrc) opts.workerSrc = '/pdf.worker.min.mjs'
-        }
-        const doc = await pdfjs.getDocument({ data: arrayBuffer }).promise
-        const parts: string[] = []
-        for (let i = 1; i <= Math.min(40, doc.numPages); i++) {
-          const page = await doc.getPage(i)
-          const tc = await page.getTextContent()
-          const pageText = (tc.items as Array<{ str?: string }>)
-            .map(it => it.str ?? '')
-            .join(' ')
-            .replace(/[ \t]+/g, ' ')
-            .trim()
-          if (pageText) parts.push(pageText)
-        }
-        return parts.join('\n\n')
-      } catch (e) {
-        console.error('Marking scheme PDF extraction failed:', e)
-        return ''
-      }
-    }
-
-    // Render an uploaded marking-scheme PDF's pages to JPEG data URLs for the
-    // vision path — used when text extraction comes back sparse (scanned or
-    // image-flattened PDFs), so image-based schemes still autopopulate.
-    const renderMarkingSchemePages = async (file: File): Promise<string[]> => {
-      const isPdf = file.type === 'application/pdf' || file.name.toLowerCase().endsWith('.pdf')
-      if (!isPdf || typeof window === 'undefined') return []
-      try {
-        const arrayBuffer = await file.arrayBuffer()
-        const pdfjs = await import('pdfjs-dist')
-        const opts = (pdfjs as { GlobalWorkerOptions?: { workerSrc?: string } }).GlobalWorkerOptions
-        if (opts && !opts.workerSrc) opts.workerSrc = '/pdf.worker.min.mjs'
-        const doc = await pdfjs.getDocument({ data: arrayBuffer }).promise
-        const pages: string[] = []
-        const max = Math.min(8, doc.numPages) // endpoint caps pdfPages at 8
-        for (let i = 1; i <= max; i++) {
-          const page = await doc.getPage(i)
-          const base = page.getViewport({ scale: 1 })
-          const scale = Math.min(2, 1400 / base.width)
-          const viewport = page.getViewport({ scale })
-          const canvas = document.createElement('canvas')
-          canvas.width = Math.ceil(viewport.width)
-          canvas.height = Math.ceil(viewport.height)
-          const ctx = canvas.getContext('2d')
-          if (!ctx) continue
-          await page.render({ canvasContext: ctx, viewport }).promise
-          pages.push(canvas.toDataURL('image/jpeg', 0.72))
-        }
-        return pages
-      } catch (e) {
-        console.error('Marking scheme PDF render failed:', e)
-        return []
-      }
-    }
-
-    // Parse an uploaded marking scheme and fill each question's answer, accepted
-    // variants, marks and marking guidance by matching question numbers. Edits
-    // apply via applyDmiEdit so they persist. Falls back to the vision path when
-    // the PDF has little extractable text (scanned schemes).
-    const handleMarkingSchemeFile = async (file: File, source: 'task' | 'assessment') => {
-      const items = source === 'task' ? taskDmiItems : assessmentDmiItems
-      if (items.length === 0) return
-      setMarkingSchemeLoading(true)
-      try {
-        toast.info('Reading the marking scheme…')
-        // Match on the paper's REAL question reference (e.g. "1(a)") — not the
-        // re-serialized questionNumber. A scheme keyed to 1(a), 1(b), 2 … never
-        // lines up with a 1, 2, 3 … serial, which left most answers unfilled.
-        // refKey (from the shared lib) normalizes "1(a)" and "1a" to the same key.
-        const questions = items.map(it => ({
-          ref: String(it.questionLabel ?? it.questionNumber ?? ''),
-          label: it.questionText,
-        }))
-        // Pass the badge's board/subject so the model leans on that board's
-        // marking conventions (it still verifies against the scheme itself).
-        const examVersions = source === 'task' ? taskDmiVersions : assessmentDmiVersions
-        const activeExamVerId = testPciViewMode.startsWith('dmi_')
-          ? testPciViewMode.slice('dmi_'.length)
-          : null
-        const activeExamVer =
-          examVersions.find(v => v.id === activeExamVerId) ?? examVersions[examVersions.length - 1]
-        const derivedExam = deriveExamContext(designatedFolder, courseName)
-        const examBody = activeExamVer?.examBody ?? derivedExam.examBody
-        const examSubject = activeExamVer?.subject ?? derivedExam.subject
-        // The tutor's PCI instructions for this context — steer which accepted
-        // forms / award rules the model favours when extracting answers. Use the
-        // active task extension's PCI when one is active, else the base PCI.
-        const pciText = (
-          source === 'task'
-            ? taskBuilder.activeExtensionId
-              ? taskBuilder.extensions.find(e => e.id === taskBuilder.activeExtensionId)?.pci
-              : taskBuilder.taskPci
-            : assessmentBuilder.taskPci
-        )?.trim()
-        const hint = { examBody, subject: examSubject, pci: pciText || undefined }
-        const content = (await extractMarkingSchemeText(file)).slice(0, 80000).trim()
-
-        let body: {
-          questions: typeof questions
-          content?: string
-          pdfPages?: string[]
-          examBody?: string
-          subject?: string
-          pci?: string
-        }
-        if (content.length >= 200) {
-          body = { questions, content, ...hint }
-        } else {
-          // Sparse text → try the vision path (scanned / image-based PDF).
-          toast.info('Scanned scheme detected — reading the pages…')
-          const pdfPages = await renderMarkingSchemePages(file)
-          if (pdfPages.length > 0) {
-            body = { questions, pdfPages, ...hint }
-          } else if (content.length >= 20) {
-            body = { questions, content, ...hint }
-          } else {
-            toast.error('Could not read the marking scheme. Upload a clearer PDF or a .txt file.')
-            return
-          }
-        }
-
-        const res = await fetch('/api/ai/parse-marking-scheme', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(body),
-        })
-        if (!res.ok) {
-          const e = await res.json().catch(() => ({}))
-          toast.error(e?.error || 'Failed to parse the marking scheme')
-          return
-        }
-        const data = await res.json()
-        // Adopt the board/subject the scheme detected — but only for fields the
-        // tutor hasn't already set on this DMI (never override a manual choice).
-        const detectPatch: { examBody?: string; subject?: string } = {}
-        if (!activeExamVer?.examBody && typeof data?.detectedExamBody === 'string') {
-          detectPatch.examBody = data.detectedExamBody
-        }
-        if (!activeExamVer?.subject && typeof data?.detectedSubject === 'string') {
-          detectPatch.subject = data.detectedSubject
-        }
-        if (detectPatch.examBody || detectPatch.subject) {
-          setExamContext(source, detectPatch)
-          toast.info(
-            `Detected ${[detectPatch.examBody, detectPatch.subject].filter(Boolean).join(' · ')} from the scheme`
-          )
-        }
-        const matches: Array<{
-          ref: string
-          answer: string
-          variants?: string[]
-          marks?: number
-          rubric?: string
-          extra?: boolean
-        }> = Array.isArray(data?.matches) ? data.matches : []
-        if (matches.length === 0) {
-          toast.error('No answers could be matched from that marking scheme.')
-          return
-        }
-        const buildPatch = (m: (typeof matches)[number]): Partial<DMIQuestion> => ({
-          answer: m.answer,
-          answerProvenance: 'answer_sheet_extracted',
-          ...(Array.isArray(m.variants) && m.variants.length > 0
-            ? { acceptableVariants: m.variants }
-            : {}),
-          ...(typeof m.marks === 'number' && m.marks > 0 ? { marks: m.marks } : {}),
-          ...(m.rubric ? { rubric: m.rubric } : {}),
-        })
-        // Split the scheme's answers into patches for EXISTING questions (keyed by
-        // reference) and EXTRA questions the scheme covers but the DMI lacks.
-        const validRefs = new Set(items.map(it => refKey(it.questionLabel ?? it.questionNumber)))
-        const patchByRef = new Map<string, Partial<DMIQuestion>>()
-        const extraMatches = matches.filter(m => m.extra && !validRefs.has(refKey(m.ref)) && m.ref)
-        for (const m of matches) {
-          if (m.extra) continue
-          const key = refKey(m.ref)
-          if (!key || !validRefs.has(key)) continue
-          patchByRef.set(key, buildPatch(m))
-        }
-        // New rows for references the scheme covers that aren't in the DMI yet —
-        // sorted naturally (3a before 3b before 10), deduped against existing refs.
-        const seenExtra = new Set<string>()
-        const newRows: DMIQuestion[] = extraMatches
-          .filter(m => {
-            const k = refKey(m.ref)
-            if (seenExtra.has(k)) return false
-            seenExtra.add(k)
-            return true
-          })
-          .sort((a, b) => a.ref.localeCompare(b.ref, undefined, { numeric: true }))
-          .map(
-            (m): DMIQuestion => ({
-              id: `dmi-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`,
-              questionNumber: 0,
-              questionLabel: m.ref,
-              questionText: `Question ${m.ref}`,
-              questionType: 'short' as DmiQuestionType,
-              ...buildPatch(m),
-              answer: m.answer,
-            })
-          )
-        // Apply the answers directly onto the questions shown in the modal (the
-        // snapshot captured above), keyed by question number — NOT via a functional
-        // update against the latest state. The items can be re-sourced (fresh ids,
-        // sometimes a whole new array) during the multi-second AI call; a functional
-        // update could then land on an array whose numbers no longer match and fill
-        // NOTHING while we still reported "Filled N". The snapshot is exactly what
-        // the tutor is looking at, so the answers always appear on those rows. We
-        // count the rows that actually changed, so the toast can't over-report.
-        const patchOnto = (arr: DMIQuestion[]) =>
-          arr.map(q => {
-            const patch = patchByRef.get(refKey(q.questionLabel ?? q.questionNumber))
-            return patch ? { ...q, ...patch } : q
-          })
-        const patchedExisting = patchOnto(items)
-        const filled = patchedExisting.reduce((n, q, i) => (q === items[i] ? n : n + 1), 0)
-        // The marking scheme edits the DMI: existing rows get answers, and any
-        // questions the scheme covers that the DMI was missing are appended.
-        const patchedItems = [...patchedExisting, ...newRows]
-        if (filled === 0 && newRows.length === 0) {
-          // Matches came back but none lined up with these questions' numbers.
-          toast.error(
-            "Couldn't line up the marking scheme with these questions — the question numbers didn't match. Check that the scheme covers the same questions."
-          )
-          return
-        }
-        // Commit to the live items (what the modal + deploy read) and the active DMI
-        // version (persisted with the course) so the answers show now AND survive a
-        // save/reload.
-        const activeVersionId = testPciViewMode.startsWith('dmi_')
-          ? testPciViewMode.slice('dmi_'.length)
-          : null
-        const patchVersions = (vs: DMIVersion[]) => {
-          if (vs.length === 0) return vs
-          const targetId = activeVersionId ?? vs[vs.length - 1].id
-          return vs.map(v =>
-            v.id === targetId ? { ...v, items: [...patchOnto(v.items), ...newRows] } : v
-          )
-        }
-        if (source === 'task') {
-          setTaskDmiItems(patchedItems)
-          setTaskDmiVersions(patchVersions)
-        } else {
-          setAssessmentDmiItems(patchedItems)
-          setAssessmentDmiVersions(patchVersions)
-        }
-        // Highlight + scroll to the freshly-appended rows so they're easy to find.
-        if (newRows.length > 0) {
-          const ids = new Set(newRows.map(r => r.id))
-          setRecentlyAddedRowIds(ids)
-          setTimeout(() => {
-            dmiRowsRef.current?.scrollTo({
-              top: dmiRowsRef.current.scrollHeight,
-              behavior: 'smooth',
-            })
-          }, 120)
-          setTimeout(() => setRecentlyAddedRowIds(new Set()), 6000)
-        }
-        const addedNote =
-          newRows.length > 0
-            ? `; added ${newRows.length} new question${newRows.length === 1 ? '' : 's'} the scheme covered`
-            : ''
-        toast.success(
-          `Filled ${filled} of ${questions.length} answers (with variants & marks)${addedNote} from the marking scheme.`
-        )
-      } catch (err) {
-        console.error('Marking scheme parse failed:', err)
-        toast.error('Failed to read the marking scheme')
-      } finally {
-        setMarkingSchemeLoading(false)
-      }
-    }
+    // Marking-scheme upload (extract → parse → fill/append) lives in its own hook.
+    const {
+      markingSchemeLoading,
+      markingSchemeInputRef,
+      recentlyAddedRowIds,
+      dmiRowsRef,
+      handleMarkingSchemeFile,
+    } = useMarkingScheme({
+      taskDmiItems,
+      assessmentDmiItems,
+      setTaskDmiItems,
+      setAssessmentDmiItems,
+      taskDmiVersions,
+      assessmentDmiVersions,
+      setTaskDmiVersions,
+      setAssessmentDmiVersions,
+      testPciViewMode,
+      taskBuilder,
+      assessmentBuilder,
+      designatedFolder,
+      courseName,
+      setExamContext,
+    })
 
     // Generate DMI using AI from content or PDF images with versioning.
     // `questionSpec` is supplied (via the spec dialog) when the source is study
