+++ tutorme-app/src/lib/assessment/marking-scheme.ts
@@ -6,6 +6,9 @@
  * isolation rather than buried in an 11k-line component.
  */
 
+import type { DMIQuestion } from '@/app/[locale]/tutor/dashboard/components/builder-types'
+import type { DmiQuestionType } from '@/lib/assessment/question-types'
+
 /** Examining bodies the marking-scheme badge can show / a tutor can pick from.
  *  "Other" labels a board we don't list yet. Order matters for derivation:
  *  more specific labels (IGCSE before GCSE) come first. */
@@ -107,4 +110,98 @@ export function deriveExamContext(
     .trim()
   if (!subject) subject = String(fallbackSubject ?? '').trim()
   return { examBody, subject: subject || undefined }
+}
+
+/** A single answer the marking-scheme parser returned for a question reference. */
+export interface SchemeMatchInput {
+  ref: string
+  answer: string
+  variants?: string[]
+  marks?: number
+  rubric?: string
+  /** True when the scheme covers a reference the DMI doesn't have yet. */
+  extra?: boolean
+}
+
+export interface AppliedScheme {
+  /** Existing items patched with their answers, then the new rows appended. */
+  patchedItems: DMIQuestion[]
+  /** Brand-new rows for references the DMI was missing (sorted, deduped). */
+  newRows: DMIQuestion[]
+  /** How many EXISTING rows actually changed (so a toast can't over-report). */
+  filled: number
+  /** Apply the same patches+appends to another array (e.g. a saved version). */
+  applyToVersionItems: (vsItems: DMIQuestion[]) => DMIQuestion[]
+}
+
+/**
+ * Pure core of the marking-scheme apply: given the DMI's current questions and the
+ * parser's matches, split into patches for existing rows (keyed by normalized
+ * reference) and brand-new rows for references the DMI lacks, and report how many
+ * existing rows changed. No React, no I/O — unit-tested in isolation.
+ *
+ * `makeId` is injected so callers control id generation (and tests stay
+ * deterministic); it defaults to a time+random id.
+ */
+export function applySchemeMatches(
+  items: DMIQuestion[],
+  matches: SchemeMatchInput[],
+  makeId: () => string = () => `dmi-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`
+): AppliedScheme {
+  const buildPatch = (m: SchemeMatchInput): Partial<DMIQuestion> => ({
+    answer: m.answer,
+    answerProvenance: 'answer_sheet_extracted',
+    ...(Array.isArray(m.variants) && m.variants.length > 0
+      ? { acceptableVariants: m.variants }
+      : {}),
+    ...(typeof m.marks === 'number' && m.marks > 0 ? { marks: m.marks } : {}),
+    ...(m.rubric ? { rubric: m.rubric } : {}),
+  })
+
+  const validRefs = new Set(items.map(it => refKey(it.questionLabel ?? it.questionNumber)))
+  const patchByRef = new Map<string, Partial<DMIQuestion>>()
+  for (const m of matches) {
+    if (m.extra) continue
+    const key = refKey(m.ref)
+    if (!key || !validRefs.has(key)) continue
+    patchByRef.set(key, buildPatch(m))
+  }
+
+  const seenExtra = new Set<string>()
+  const newRows: DMIQuestion[] = matches
+    .filter(m => m.extra && !validRefs.has(refKey(m.ref)) && m.ref)
+    .filter(m => {
+      const k = refKey(m.ref)
+      if (seenExtra.has(k)) return false
+      seenExtra.add(k)
+      return true
+    })
+    .sort((a, b) => a.ref.localeCompare(b.ref, undefined, { numeric: true }))
+    .map(
+      (m): DMIQuestion => ({
+        id: makeId(),
+        questionNumber: 0,
+        questionLabel: m.ref,
+        questionText: `Question ${m.ref}`,
+        questionType: 'short' as DmiQuestionType,
+        ...buildPatch(m),
+        answer: m.answer,
+      })
+    )
+
+  const patchOnto = (arr: DMIQuestion[]) =>
+    arr.map(q => {
+      const patch = patchByRef.get(refKey(q.questionLabel ?? q.questionNumber))
+      return patch ? { ...q, ...patch } : q
+    })
+
+  const patchedExisting = patchOnto(items)
+  const filled = patchedExisting.reduce((n, q, i) => (q === items[i] ? n : n + 1), 0)
+
+  return {
+    patchedItems: [...patchedExisting, ...newRows],
+    newRows,
+    filled,
+    applyToVersionItems: (vsItems: DMIQuestion[]) => [...patchOnto(vsItems), ...newRows],
+  }
 }
